requirements.txt
fastparquet~=2024.2.0
gym~=0.26.2
numpy~=1.26.4
matplotlib~=3.8.4
matplotlib-inline~=0.1.7
opencv-python~=4.9.0.80
optuna~=3.6.1
optuna-dashboard~=0.16.0
pandas~=2.2.2
pillow~=10.3.0
pyarrow~=16.0.0
pyaml~=6.0.1
scipy~=1.13.0
statsmodels~=0.14.4
scikit-learn~=1.5.1
torch~=2.3.1
torchvision~=0.18.1
